Extract targeting & camera basis utilities
Move and centralize targeting and camera math logic into reusable classes. Added FocusTargetSelector to encapsulate target finding, filtering, directional swap, and sight checks (extracted from LockOnHandler). Introduced FocusCameraBasisUtil for camera basis/rotation utilities and reused it in FocusShoulderSurfingCameraSystem and FocusCameraOcclusionEvaluator. Added FocusCameraMath.targetSwapBlendToNormal and updated callers (FocusCameraController, FocusDefaultCameraRotationFollowPolicy) to remove duplicated easing code. Refactored LockOnHandler to use the new selector, simplify lock/unlock and grace-timer logic, and improve camera enforcement and swap cooldown handling. Updated HUD overlay (FocusCorrectedCrosshairOverlay) to use FocusScreenProjectionUtil for screen projection. Miscellaneous small cleanup and method extraction to reduce duplication and improve clarity.
