+++ src/main/java/com/jvn/focus/client/LockOnHandler.java
@@ -3,24 +3,12 @@ package com.jvn.focus.client;
 import com.jvn.focus.Focus;
 import com.jvn.focus.client.camera.FocusCameraController;
 import com.jvn.focus.client.camera.FocusCameraPose;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
 import net.minecraft.client.CameraType;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
-import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.Mob;
-import net.minecraft.world.entity.NeutralMob;
-import net.minecraft.world.entity.monster.Enemy;
-import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.level.ClipContext;
-import net.minecraft.world.phys.BlockHitResult;
-import net.minecraft.world.phys.HitResult;
 import net.minecraft.world.phys.Vec2;
 import net.minecraft.world.phys.Vec3;
 import net.neoforged.api.distmarker.Dist;
@@ -33,11 +21,6 @@ import net.neoforged.neoforge.client.event.RenderFrameEvent;
 
 @EventBusSubscriber(modid = Focus.MOD_ID, value = Dist.CLIENT)
 public final class LockOnHandler {
-    private static final double MAX_LOCK_DISTANCE = 20.0D;
-    private static final double MAX_LOCK_DISTANCE_SQR = MAX_LOCK_DISTANCE * MAX_LOCK_DISTANCE;
-    private static final double OCCLUDED_LOCK_DISTANCE = 10.0D;
-    private static final double OCCLUDED_LOCK_DISTANCE_SQR = OCCLUDED_LOCK_DISTANCE * OCCLUDED_LOCK_DISTANCE;
-    private static final double LOCK_ON_FOV_THRESHOLD = 0.35D;
     private static final int TARGET_SWAP_MIN_COOLDOWN_TICKS = 8;
     private static final int OCCLUDED_GRACE_TICKS = 15;
     private static final int OUT_OF_RANGE_GRACE_TICKS = 10;
@@ -85,52 +68,68 @@ public final class LockOnHandler {
         CAMERA_CONTROLLER.setFreeLookInputActive(freeLookActive);
         handleCameraAdjustmentInput(player);
 
-        if (lockedTarget != null && !lockedTarget.isAlive()) {
-            LivingEntity replacementTarget = Targeting.findClosestTarget(player, lockedTarget);
+        updateLockOnState(player, minecraft);
+        CAMERA_CONTROLLER.updatePlayerVisibility(player, lockedTarget, 1.0F);
+        updateTargetSwapCooldown();
+        CAMERA_CONTROLLER.onClientTick(lockedTarget != null);
+        if (lockedTarget != null && !freeLookActive) {
+            tryDirectionalTargetSwap(player);
+        } else {
+            resetTargetSwapInput();
+        }
+        enforceCameraType(minecraft);
+    }
+
+    private static void updateLockOnState(LocalPlayer player, Minecraft minecraft) {
+        if (lockedTarget == null) {
+            return;
+        }
+
+        if (!lockedTarget.isAlive()) {
+            LivingEntity replacementTarget = FocusTargetSelector.findClosestTarget(player, lockedTarget);
             if (replacementTarget != null) {
                 resetTargetSwapInput();
                 setLockedTarget(player, replacementTarget, true);
                 occlusionGraceTicks = 0;
                 outOfRangeGraceTicks = 0;
             } else {
-                lockedTarget = null;
-                restoreCamera(minecraft);
-                player.displayClientMessage(Component.translatable("message.focus.lock_on.lost"), true);
+                unlockWithMessage(player, minecraft, "message.focus.lock_on.lost");
             }
-        } else if (lockedTarget != null && (isLockOnHiddenFromPlayer(player, lockedTarget) || !Targeting.isTargetAllowed(lockedTarget))) {
-            lockedTarget = null;
-            restoreCamera(minecraft);
-            player.displayClientMessage(Component.translatable("message.focus.lock_on.lost"), true);
-        } else if (lockedTarget != null) {
-            boolean obstructed = !hasDirectSight(player, lockedTarget);
-            double distanceSqr = player.distanceToSqr(lockedTarget);
-            double maxDistanceSqr = obstructed ? OCCLUDED_LOCK_DISTANCE_SQR : MAX_LOCK_DISTANCE_SQR;
-
-            if (distanceSqr > maxDistanceSqr) {
-                outOfRangeGraceTicks++;
-            } else {
-                outOfRangeGraceTicks = 0;
-            }
-            if (obstructed) {
-                occlusionGraceTicks++;
-            } else {
-                occlusionGraceTicks = 0;
-            }
-
-            if (outOfRangeGraceTicks > OUT_OF_RANGE_GRACE_TICKS) {
-                lockedTarget = null;
-                restoreCamera(minecraft);
-                player.displayClientMessage(
-                        Component.translatable(obstructed ? "message.focus.lock_on.obstructed" : "message.focus.lock_on.lost"),
-                        true);
-            } else if (occlusionGraceTicks > OCCLUDED_GRACE_TICKS) {
-                lockedTarget = null;
-                restoreCamera(minecraft);
-                player.displayClientMessage(Component.translatable("message.focus.lock_on.obstructed"), true);
-            }
+            return;
         }
-        CAMERA_CONTROLLER.updatePlayerVisibility(player, lockedTarget, 1.0F);
 
+        if (FocusTargetSelector.isLockOnHiddenFromPlayer(player, lockedTarget) || !FocusTargetSelector.isTargetAllowed(lockedTarget)) {
+            unlockWithMessage(player, minecraft, "message.focus.lock_on.lost");
+            return;
+        }
+
+        updateGraceTimers(player, minecraft);
+    }
+
+    private static void updateGraceTimers(LocalPlayer player, Minecraft minecraft) {
+        boolean obstructed = !FocusTargetSelector.hasDirectSight(player, lockedTarget);
+        double distanceSqr = player.distanceToSqr(lockedTarget);
+        double maxDistanceSqr = obstructed
+                ? FocusTargetSelector.OCCLUDED_LOCK_DISTANCE_SQR
+                : FocusTargetSelector.MAX_LOCK_DISTANCE_SQR;
+
+        outOfRangeGraceTicks = distanceSqr > maxDistanceSqr ? outOfRangeGraceTicks + 1 : 0;
+        occlusionGraceTicks = obstructed ? occlusionGraceTicks + 1 : 0;
+
+        if (outOfRangeGraceTicks > OUT_OF_RANGE_GRACE_TICKS) {
+            unlockWithMessage(player, minecraft, obstructed ? "message.focus.lock_on.obstructed" : "message.focus.lock_on.lost");
+        } else if (occlusionGraceTicks > OCCLUDED_GRACE_TICKS) {
+            unlockWithMessage(player, minecraft, "message.focus.lock_on.obstructed");
+        }
+    }
+
+    private static void unlockWithMessage(LocalPlayer player, Minecraft minecraft, String messageKey) {
+        lockedTarget = null;
+        restoreCamera(minecraft);
+        player.displayClientMessage(Component.translatable(messageKey), true);
+    }
+
+    private static void updateTargetSwapCooldown() {
         if (targetSwapCooldownTicks > 0) {
             targetSwapCooldownTicks--;
             if (targetSwapCooldownTicks == 0 && !targetSwapReadyForNewFlick) {
@@ -139,16 +138,11 @@ public final class LockOnHandler {
                 pendingMouseDeltaY = 0.0D;
             }
         }
-        CAMERA_CONTROLLER.onClientTick(lockedTarget != null);
-        if (lockedTarget != null && !freeLookActive) {
-            tryDirectionalTargetSwap(player);
-        } else {
-            resetTargetSwapInput();
-        }
+    }
 
+    private static void enforceCameraType(Minecraft minecraft) {
         if (lockedTarget != null) {
             CameraType cameraType = minecraft.options.getCameraType();
-            // Detect user-initiated perspective changes (F5) and remember their preference
             boolean userChangedPerspective = lastEnforcedCameraType != null && cameraType != lastEnforcedCameraType;
             boolean allowFirstPerson = FocusClientConfig.allowFirstPersonWhileTargeting();
             boolean allowFrontFacing = FocusClientConfig.allowFrontFacingThirdPersonWhileTargeting();
@@ -240,7 +234,7 @@ public final class LockOnHandler {
             return;
         }
 
-        LivingEntity nextTarget = findTarget(player);
+        LivingEntity nextTarget = FocusTargetSelector.findTarget(player);
         if (nextTarget == null) {
             player.displayClientMessage(Component.translatable("message.focus.lock_on.no_target"), true);
             return;
@@ -285,18 +279,6 @@ public final class LockOnHandler {
     private static void setLockedTarget(LocalPlayer player, LivingEntity nextTarget, boolean applySwapSmoothing) {
         lockedTarget = nextTarget;
         CAMERA_CONTROLLER.onTargetSet(player, nextTarget, applySwapSmoothing);
-    }
-
-    private static LivingEntity findTarget(LocalPlayer player) {
-        return Targeting.findTarget(player);
-    }
-
-    private static boolean hasDirectSight(LocalPlayer player, LivingEntity target) {
-        return Targeting.hasDirectSight(player, target);
-    }
-
-    private static boolean isLockOnHiddenFromPlayer(LocalPlayer player, LivingEntity target) {
-        return Targeting.isLockOnHiddenFromPlayer(player, target);
     }
 
     public static void onRawMouseInput(double deltaX, double deltaY) {
@@ -361,7 +343,7 @@ public final class LockOnHandler {
     }
 
     public static boolean hasLineOfSightToLockedTarget(LocalPlayer player) {
-        return lockedTarget != null && player != null && hasDirectSight(player, lockedTarget);
+        return lockedTarget != null && player != null && FocusTargetSelector.hasDirectSight(player, lockedTarget);
     }
 
     public static boolean isLockedTargetWithinHitRange(LocalPlayer player) {
@@ -431,7 +413,7 @@ public final class LockOnHandler {
 
         Vec2 mouseDirection = new Vec2((float) pendingMouseDeltaX, (float) -pendingMouseDeltaY);
         Vec3 cameraLookDir = CAMERA_CONTROLLER.getSmoothedLookDirection();
-        LivingEntity swappedTarget = Targeting.findDirectionalTarget(player, lockedTarget, mouseDirection, cameraLookDir);
+        LivingEntity swappedTarget = FocusTargetSelector.findDirectionalTarget(player, lockedTarget, mouseDirection, cameraLookDir);
         if (swappedTarget == null) {
             dampenTargetSwapInput(0.45D);
             return;
@@ -518,289 +500,4 @@ public final class LockOnHandler {
         }
     }
 
-    private static final class Targeting {
-        private Targeting() {}
-
-        private static LivingEntity findTarget(LocalPlayer player) {
-            Vec3 eyePosition = player.getEyePosition();
-            Vec3 lookDirection = player.getLookAngle().normalize();
-            TargetFilterSettings targetFilterSettings = readTargetFilterSettings();
-            LivingEntity bestTarget = null;
-            double bestAlignment = LOCK_ON_FOV_THRESHOLD;
-            double bestDistanceSqr = Double.MAX_VALUE;
-
-            for (LivingEntity entity : player.level().getEntitiesOfClass(
-                    LivingEntity.class,
-                    player.getBoundingBox().inflate(MAX_LOCK_DISTANCE),
-                    candidate -> candidate != player && candidate.isAlive() && isTargetAllowed(candidate, targetFilterSettings))) {
-                if (isLockOnHiddenFromPlayer(player, entity) || !isWithinAllowedLockDistance(player, entity)) {
-                    continue;
-                }
-
-                double alignment = getTargetAlignment(eyePosition, lookDirection, entity);
-                if (alignment < LOCK_ON_FOV_THRESHOLD) {
-                    continue;
-                }
-
-                double distanceSqr = player.distanceToSqr(entity);
-                if (alignment > bestAlignment + 1.0E-4D
-                        || (Math.abs(alignment - bestAlignment) <= 1.0E-4D && distanceSqr < bestDistanceSqr)) {
-                    bestAlignment = alignment;
-                    bestDistanceSqr = distanceSqr;
-                    bestTarget = entity;
-                }
-            }
-
-            return bestTarget;
-        }
-
-        private static LivingEntity findClosestTarget(LocalPlayer player, LivingEntity excludedTarget) {
-            TargetFilterSettings targetFilterSettings = readTargetFilterSettings();
-            LivingEntity bestTarget = null;
-            double bestDistanceSqr = Double.MAX_VALUE;
-
-            for (LivingEntity entity : player.level().getEntitiesOfClass(
-                    LivingEntity.class,
-                    player.getBoundingBox().inflate(MAX_LOCK_DISTANCE),
-                    candidate -> candidate != player
-                            && candidate.isAlive()
-                            && candidate != excludedTarget
-                            && isTargetAllowed(candidate, targetFilterSettings))) {
-                if (isLockOnHiddenFromPlayer(player, entity) || !isWithinAllowedLockDistance(player, entity)) {
-                    continue;
-                }
-
-                double distanceSqr = player.distanceToSqr(entity);
-                if (distanceSqr < bestDistanceSqr) {
-                    bestDistanceSqr = distanceSqr;
-                    bestTarget = entity;
-                }
-            }
-
-            return bestTarget;
-        }
-
-        private static LivingEntity findDirectionalTarget(LocalPlayer player, LivingEntity currentTarget, Vec2 mouseDirection, Vec3 cameraLookDirection) {
-            if (currentTarget == null) {
-                return null;
-            }
-
-            float mouseMagnitude = Mth.sqrt(
-                    mouseDirection.x * mouseDirection.x + mouseDirection.y * mouseDirection.y);
-            if (mouseMagnitude < 1.0E-4F) {
-                return null;
-            }
-
-            float mouseDirX = mouseDirection.x / mouseMagnitude;
-            float mouseDirY = mouseDirection.y / mouseMagnitude;
-            Vec3 eyePosition = player.getEyePosition();
-            // Use camera look direction for accurate screen-space projection during lock-on
-            Vec3 lookDirection = cameraLookDirection.lengthSqr() > 1.0E-6D
-                    ? cameraLookDirection.normalize()
-                    : player.getLookAngle().normalize();
-            Vec3 right = lookDirection.cross(new Vec3(0.0D, 1.0D, 0.0D));
-            if (right.lengthSqr() < 1.0E-6D) {
-                right = new Vec3(1.0D, 0.0D, 0.0D);
-            } else {
-                right = right.normalize();
-            }
-            Vec3 up = right.cross(lookDirection).normalize();
-
-            Vec2 currentTargetScreen = projectToScreenSpace(eyePosition, lookDirection, right, up, currentTarget);
-            if (currentTargetScreen == null) {
-                return null;
-            }
-
-            double directionThreshold = FocusClientConfig.targetSwapDirectionThreshold();
-            double minScreenSeparation = FocusClientConfig.targetSwapMinScreenSeparation();
-            TargetFilterSettings targetFilterSettings = readTargetFilterSettings();
-            LivingEntity bestTarget = null;
-            float bestScreenDistance = Float.MAX_VALUE;
-            float bestAlignment = (float) directionThreshold;
-            double bestDistanceSqr = Double.MAX_VALUE;
-
-            for (LivingEntity entity : player.level().getEntitiesOfClass(
-                    LivingEntity.class,
-                    player.getBoundingBox().inflate(MAX_LOCK_DISTANCE),
-                    candidate -> candidate != player
-                            && candidate.isAlive()
-                            && candidate != currentTarget
-                            && isTargetAllowed(candidate, targetFilterSettings))) {
-                if (isLockOnHiddenFromPlayer(player, entity) || !isWithinAllowedLockDistance(player, entity)) {
-                    continue;
-                }
-
-                Vec2 candidateScreen = projectToScreenSpace(eyePosition, lookDirection, right, up, entity);
-                if (candidateScreen == null) {
-                    continue;
-                }
-
-                float deltaX = candidateScreen.x - currentTargetScreen.x;
-                float deltaY = candidateScreen.y - currentTargetScreen.y;
-                float candidateScreenDistance = Mth.sqrt(deltaX * deltaX + deltaY * deltaY);
-                if (candidateScreenDistance < minScreenSeparation) {
-                    continue;
-                }
-
-                float candidateDirX = deltaX / candidateScreenDistance;
-                float candidateDirY = deltaY / candidateScreenDistance;
-                float directionalAlignment = candidateDirX * mouseDirX + candidateDirY * mouseDirY;
-                if (directionalAlignment < directionThreshold) {
-                    continue;
-                }
-
-                double distanceSqr = player.distanceToSqr(entity);
-                if (candidateScreenDistance < bestScreenDistance - 1.0E-4F
-                        || (Math.abs(candidateScreenDistance - bestScreenDistance) <= 1.0E-4F
-                                && directionalAlignment > bestAlignment + 1.0E-4F)
-                        || (Math.abs(candidateScreenDistance - bestScreenDistance) <= 1.0E-4F
-                                && Math.abs(directionalAlignment - bestAlignment) <= 1.0E-4F
-                                && distanceSqr < bestDistanceSqr)) {
-                    bestScreenDistance = candidateScreenDistance;
-                    bestAlignment = directionalAlignment;
-                    bestDistanceSqr = distanceSqr;
-                    bestTarget = entity;
-                }
-            }
-
-            return bestTarget;
-        }
-
-        private static Vec3 getTargetAimPoint(LivingEntity target, float partialTick) {
-            return CAMERA_CONTROLLER.targetPointFor(target, partialTick);
-        }
-
-        private static double getTargetAlignment(Vec3 eyePosition, Vec3 lookDirection, LivingEntity target) {
-            Vec3 toTarget = getTargetAimPoint(target, 1.0F).subtract(eyePosition);
-            double lengthSqr = toTarget.lengthSqr();
-            if (lengthSqr <= 1.0E-6D) {
-                return -1.0D;
-            }
-
-            return toTarget.scale(1.0D / Math.sqrt(lengthSqr)).dot(lookDirection);
-        }
-
-        private static Vec2 projectToScreenSpace(
-                Vec3 eyePosition,
-                Vec3 lookDirection,
-                Vec3 right,
-                Vec3 up,
-                LivingEntity target) {
-            Vec3 toTarget = getTargetAimPoint(target, 1.0F).subtract(eyePosition);
-            double forward = toTarget.dot(lookDirection);
-            if (forward <= 1.0E-3D) {
-                return null;
-            }
-
-            double screenX = toTarget.dot(right) / forward;
-            double screenY = toTarget.dot(up) / forward;
-            return new Vec2((float) screenX, (float) screenY);
-        }
-
-        private static boolean isWithinAllowedLockDistance(LocalPlayer player, LivingEntity target) {
-            boolean obstructed = !hasDirectSight(player, target);
-            double maxDistanceSqr = obstructed ? OCCLUDED_LOCK_DISTANCE_SQR : MAX_LOCK_DISTANCE_SQR;
-            return player.distanceToSqr(target) <= maxDistanceSqr;
-        }
-
-        private static boolean hasDirectSight(LocalPlayer player, LivingEntity target) {
-            Vec3 from = player.getEyePosition();
-            Vec3 to = getTargetAimPoint(target, 1.0F);
-            BlockHitResult hitResult = player.level().clip(new ClipContext(from, to, ClipContext.Block.COLLIDER, ClipContext.Fluid.NONE, player));
-            return hitResult.getType() != HitResult.Type.BLOCK;
-        }
-
-        private static boolean isLockOnHiddenFromPlayer(LocalPlayer player, LivingEntity target) {
-            return target.isInvisible() || target.isInvisibleTo(player);
-        }
-
-        private static boolean isTargetAllowed(LivingEntity target) {
-            return isTargetAllowed(target, readTargetFilterSettings());
-        }
-
-        private static boolean isTargetAllowed(LivingEntity target, TargetFilterSettings targetFilterSettings) {
-            if (!targetFilterSettings.enabled()) {
-                return true;
-            }
-
-            boolean matchesAnyFilter = matchesAnyFilter(target, targetFilterSettings);
-            if (targetFilterSettings.mode() == FocusClientConfig.TargetFilterMode.EXCLUSIVE) {
-                return matchesAnyFilter;
-            }
-            return !matchesAnyFilter;
-        }
-
-        private static boolean matchesAnyFilter(LivingEntity target, TargetFilterSettings targetFilterSettings) {
-            if (targetFilterSettings.filterPlayers() && target instanceof Player) {
-                return true;
-            }
-            if (targetFilterSettings.filterHostileMobs() && target instanceof Enemy) {
-                return true;
-            }
-            if (targetFilterSettings.filterNeutralMobs() && target instanceof NeutralMob && !(target instanceof Enemy)) {
-                return true;
-            }
-            if (targetFilterSettings.filterPassiveMobs() && isPassiveMob(target)) {
-                return true;
-            }
-            if (!targetFilterSettings.entityTypeIds().isEmpty()) {
-                ResourceLocation entityTypeId = BuiltInRegistries.ENTITY_TYPE.getKey(target.getType());
-                if (entityTypeId != null && targetFilterSettings.entityTypeIds().contains(entityTypeId)) {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        private static boolean isPassiveMob(LivingEntity target) {
-            return target instanceof Mob && !(target instanceof Enemy) && !(target instanceof NeutralMob);
-        }
-
-        private static TargetFilterSettings readTargetFilterSettings() {
-            if (!FocusClientConfig.enableTargetFilters()) {
-                return new TargetFilterSettings(
-                        false,
-                        FocusClientConfig.TargetFilterMode.EXCLUDE,
-                        false,
-                        false,
-                        false,
-                        false,
-                        Set.of());
-            }
-
-            return new TargetFilterSettings(
-                    true,
-                    FocusClientConfig.targetFilterMode(),
-                    FocusClientConfig.filterPlayers(),
-                    FocusClientConfig.filterPassiveMobs(),
-                    FocusClientConfig.filterNeutralMobs(),
-                    FocusClientConfig.filterHostileMobs(),
-                    parseEntityTypeIds(FocusClientConfig.targetFilterEntityIds()));
-        }
-
-        private static Set<ResourceLocation> parseEntityTypeIds(List<String> configuredEntityTypeIds) {
-            if (configuredEntityTypeIds.isEmpty()) {
-                return Set.of();
-            }
-
-            Set<ResourceLocation> parsedEntityTypeIds = new HashSet<>();
-            for (String configuredEntityTypeId : configuredEntityTypeIds) {
-                ResourceLocation parsedId = ResourceLocation.tryParse(configuredEntityTypeId);
-                if (parsedId != null) {
-                    parsedEntityTypeIds.add(parsedId);
-                }
-            }
-            return parsedEntityTypeIds;
-        }
-
-        private record TargetFilterSettings(
-                boolean enabled,
-                FocusClientConfig.TargetFilterMode mode,
-                boolean filterPlayers,
-                boolean filterPassiveMobs,
-                boolean filterNeutralMobs,
-                boolean filterHostileMobs,
-                Set<ResourceLocation> entityTypeIds) {
-        }
-    }
 }
