+++ src/main/java/com/jvn/focus/client/hud/LockOnIndicatorHudOverlay.java
@@ -5,7 +5,6 @@ import com.jvn.focus.client.FocusClientConfig;
 import com.jvn.focus.client.LockOnHandler;
 import com.mojang.blaze3d.systems.RenderSystem;
 
-import net.minecraft.client.Camera;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
@@ -15,17 +14,14 @@ import net.minecraft.world.phys.Vec3;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.neoforge.client.ClientHooks;
 import net.neoforged.neoforge.client.event.RegisterGuiLayersEvent;
 import net.neoforged.neoforge.client.gui.VanillaGuiLayers;
-import org.joml.Vector3f;
 
 @SuppressWarnings("removal")
 @EventBusSubscriber(modid = Focus.MOD_ID, bus = EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
 public final class LockOnIndicatorHudOverlay {
     private static final ResourceLocation INDICATOR_LAYER = ResourceLocation.fromNamespaceAndPath(Focus.MOD_ID, "lock_on_indicator");
     private static final int SOURCE_TEXTURE_SIZE = 32;
-    private static final double MIN_DEPTH = 0.01D;
     private static final float TARGET_HEIGHT_FACTOR = 0.75F;
 
     private LockOnIndicatorHudOverlay() {}
@@ -48,7 +44,7 @@ public final class LockOnIndicatorHudOverlay {
 
         float partialTick = deltaTracker.getGameTimeDeltaPartialTick(minecraft.level.tickRateManager().runsNormally());
         Vec3 targetPoint = target.getPosition(partialTick).add(0.0D, target.getBbHeight() * TARGET_HEIGHT_FACTOR, 0.0D);
-        ScreenPoint projectedPoint = projectToScreen(minecraft, targetPoint, partialTick, guiGraphics.guiWidth(), guiGraphics.guiHeight());
+        FocusScreenProjectionUtil.ScreenPoint projectedPoint = FocusScreenProjectionUtil.projectToScreen(minecraft, targetPoint, partialTick, guiGraphics.guiWidth(), guiGraphics.guiHeight());
         if (projectedPoint == null) {
             return;
         }
@@ -60,8 +56,8 @@ public final class LockOnIndicatorHudOverlay {
             RenderSystem.defaultBlendFunc();
             LockOnIndicatorAnimationUtil.renderOotTriangleOrbit(
                     guiGraphics,
-                    projectedPoint.x,
-                    projectedPoint.y,
+                    projectedPoint.x(),
+                    projectedPoint.y(),
                     style.texture(),
                     style.drawSize(),
                     animationTicks,
@@ -72,9 +68,10 @@ public final class LockOnIndicatorHudOverlay {
         drawCenteredIndicator(guiGraphics, projectedPoint, style);
     }
 
+    @SuppressWarnings("unused")
     private static void drawCenteredIndicator(
             GuiGraphics guiGraphics,
-            ScreenPoint center,
+            FocusScreenProjectionUtil.ScreenPoint center,
             FocusClientConfig.LockOnIndicatorStyle style) {
         ResourceLocation texture = style.texture();
         int drawSize = style.drawSize();
@@ -82,7 +79,7 @@ public final class LockOnIndicatorHudOverlay {
         RenderSystem.enableBlend();
         RenderSystem.defaultBlendFunc();
         guiGraphics.pose().pushPose();
-        guiGraphics.pose().translate(center.x - halfSize, center.y - halfSize, 0.0F);
+        guiGraphics.pose().translate(center.x() - halfSize, center.y() - halfSize, 0.0F);
         guiGraphics.blit(
                 texture,
                 0,
@@ -98,44 +95,4 @@ public final class LockOnIndicatorHudOverlay {
         guiGraphics.pose().popPose();
         RenderSystem.disableBlend();
     }
-
-    private static ScreenPoint projectToScreen(Minecraft minecraft, Vec3 worldPos, float partialTick, int guiWidth, int guiHeight) {
-        Camera camera = minecraft.gameRenderer.getMainCamera();
-        Vec3 relativePos = worldPos.subtract(camera.getPosition());
-
-        Vector3f leftVector = camera.getLeftVector();
-        Vector3f upVector = camera.getUpVector();
-        Vector3f lookVector = camera.getLookVector();
-
-        double depth = dot(relativePos, lookVector);
-        if (depth <= MIN_DEPTH) {
-            return null;
-        }
-
-        double horizontal = -dot(relativePos, leftVector);
-        double vertical = dot(relativePos, upVector);
-        double configuredFov = minecraft.options.fov().get();
-        double fovDegrees = ClientHooks.getFieldOfView(minecraft.gameRenderer, camera, partialTick, configuredFov, true);
-        double tanHalfFov = Math.tan(Math.toRadians(fovDegrees * 0.5D));
-        if (tanHalfFov <= 0.0D) {
-            return null;
-        }
-
-        double aspectRatio = (double) guiWidth / (double) guiHeight;
-        double ndcX = horizontal / (depth * tanHalfFov * aspectRatio);
-        double ndcY = vertical / (depth * tanHalfFov);
-        if (Math.abs(ndcX) > 1.0D || Math.abs(ndcY) > 1.0D) {
-            return null;
-        }
-
-        float screenX = (float) ((ndcX + 1.0D) * 0.5D * guiWidth);
-        float screenY = (float) ((1.0D - ndcY) * 0.5D * guiHeight);
-        return new ScreenPoint(screenX, screenY);
-    }
-
-    private static double dot(Vec3 vector, Vector3f other) {
-        return vector.x * other.x() + vector.y * other.y() + vector.z * other.z();
-    }
-
-    private record ScreenPoint(float x, float y) {}
 }
